requirements.txt
requests==2.32.4
python-dotenv==1.0.1
linkup-sdk
uvicorn
fastapi
pydantic
openai
httpx